Move local controller resolve mthod to modelcmd
