+++ cmd/juju/commands/switch.go
@@ -99,9 +99,9 @@ func (c *switchCommand) Run(ctx *cmd.Context) (resultErr error) {
 		return errors.Errorf("cannot switch when JUJU_MODEL is overriding the model (set to %q)", model)
 	}
 
-	// If the name identifies a controller, then set that as the current one.
+	// If the target identifies a controller, then set that as the current controller.
 	var newControllerName string
-	if newControllerName, _, err = jujuclient.LocalControllerByName(c.Store, c.Target); err == nil {
+	if newControllerName, err = modelcmd.ResolveControllerByName(c.Store, c.Target); err == nil {
 		if newControllerName == currentControllerName {
 			newName = currentName
 			return nil
@@ -123,7 +123,7 @@ func (c *switchCommand) Run(ctx *cmd.Context) (resultErr error) {
 	newControllerName, modelName := modelcmd.SplitModelName(c.Target)
 	if newControllerName != "" {
 		// A controller was specified so see if we should use a local version.
-		newControllerName, _, err = jujuclient.LocalControllerByName(c.Store, newControllerName)
+		newControllerName, err = modelcmd.ResolveControllerByName(c.Store, newControllerName)
 		if err == nil {
 			newName = modelcmd.JoinModelName(newControllerName, modelName)
 		} else {
