+++ cmd/modelcmd/controller_test.go
@@ -6,6 +6,7 @@ package modelcmd_test
 import (
 	"github.com/juju/cmd"
 	"github.com/juju/cmd/cmdtesting"
+	"github.com/juju/errors"
 	jc "github.com/juju/testing/checkers"
 	gc "gopkg.in/check.v1"
 
@@ -80,4 +81,47 @@ func testEnsureControllerName(c *gc.C, store jujuclient.ClientStore, expect stri
 	cmd, err := initTestControllerCommand(c, store, args...)
 	c.Assert(err, jc.ErrorIsNil)
 	c.Assert(cmd.ControllerName(), gc.Equals, expect)
+}
+
+type ControllerSuite struct {
+	store jujuclient.ControllerStore
+}
+
+var _ = gc.Suite(&ControllerSuite{})
+
+func (s *ControllerSuite) SetUpTest(c *gc.C) {
+	controller := jujuclient.ControllerDetails{ControllerUUID: "controller-uuid"}
+	anothercontroller := jujuclient.ControllerDetails{ControllerUUID: "another-uuid"}
+	s.store = &jujuclienttesting.MemStore{
+		Controllers: map[string]jujuclient.ControllerDetails{
+			"local.controller":        controller,
+			"anothercontroller":       anothercontroller,
+			"local.anothercontroller": jujuclient.ControllerDetails{},
+		},
+	}
+}
+
+func (s *ControllerSuite) TestLocalNameFound(c *gc.C) {
+	name, err := modelcmd.ResolveControllerByName(s.store, "local.controller")
+	c.Assert(err, jc.ErrorIsNil)
+	c.Assert(name, gc.DeepEquals, "local.controller")
+}
+
+func (s *ControllerSuite) TestLocalNameFallback(c *gc.C) {
+	name, err := modelcmd.ResolveControllerByName(s.store, "controller")
+	c.Assert(name, gc.DeepEquals, "local.controller")
+	c.Assert(err, jc.ErrorIsNil)
+}
+
+func (s *ControllerSuite) TestNonLocalController(c *gc.C) {
+	name, err := modelcmd.ResolveControllerByName(s.store, "anothercontroller")
+	c.Assert(name, gc.DeepEquals, "anothercontroller")
+	c.Assert(err, jc.ErrorIsNil)
+}
+
+func (s *ControllerSuite) TestNotFound(c *gc.C) {
+	_, err := modelcmd.ResolveControllerByName(s.store, "foo")
+	c.Assert(err, jc.Satisfies, errors.IsNotFound)
+	// We should report on the passed in controller name.
+	c.Assert(err, gc.ErrorMatches, ".* foo .*")
 }
